services/analytics.py
"""
Google Analytics 4 Manager
Handles all GA4 API operations
"""
import logging
from typing import List, Dict, Any, Optional
from google.analytics.admin import AnalyticsAdminServiceClient
from google.analytics.data_v1beta import BetaAnalyticsDataClient
from google.analytics.data_v1beta.types import (
    RunReportRequest,
    DateRange,
    Dimension,
    Metric
)
from azure.identity import DefaultAzureCredential
from azure.keyvault.secrets import SecretClient
from google.oauth2 import service_account
import json
import os

logger = logging.getLogger(__name__)


class AnalyticsManager:
    """Manager for Google Analytics 4 operations"""
    
    def __init__(self):
        """Initialize Analytics Manager with credentials from Key Vault"""
        self.admin_client = None
        self.data_client = None
        self._initialize_clients()
    
    def _initialize_clients(self):
        """Initialize Google Analytics clients with service account credentials"""
        try:
            # Get credentials from Azure Key Vault
            vault_url = os.environ.get("AZURE_KEY_VAULT_URL", "https://kv-terprint-dev.vault.azure.net")
            credential = DefaultAzureCredential()
            kv_client = SecretClient(vault_url=vault_url, credential=credential)
            
            # Retrieve service account key
            secret = kv_client.get_secret("google-service-account-key")
            service_account_info = json.loads(secret.value)
            
            # Create credentials
            credentials = service_account.Credentials.from_service_account_info(
                service_account_info,
                scopes=[
                    "https://www.googleapis.com/auth/analytics.readonly",
                    "https://www.googleapis.com/auth/analytics.edit"
                ]
            )
            
            # Initialize clients
            self.admin_client = AnalyticsAdminServiceClient(credentials=credentials)
            self.data_client = BetaAnalyticsDataClient(credentials=credentials)
            
            logger.info("✅ Google Analytics clients initialized successfully")
            
        except Exception as e:
            logger.error(f"Failed to initialize Analytics clients: {e}")
            # Don't raise - allow app to start in degraded mode
    
    async def list_properties(self) -> List[Dict[str, Any]]:
        """List all GA4 properties"""
        try:
            if not self.admin_client:
                raise Exception("Analytics client not initialized")
            
            # TODO: Implement actual listing logic
            # For now, return sample data
            logger.info("Listing GA4 properties")
            return [
                {
                    "propertyId": "G-SAMPLE123",
                    "name": "Sample Property",
                    "createTime": "2026-01-01T00:00:00Z"
                }
            ]
            
        except Exception as e:
            logger.error(f"Error listing properties: {e}")
            raise
    
    async def get_property(self, property_id: str) -> Dict[str, Any]:
        """Get details for a specific GA4 property"""
        try:
            if not self.admin_client:
                raise Exception("Analytics client not initialized")
            
            logger.info(f"Getting property: {property_id}")
            
            # TODO: Implement actual get logic
            return {
                "propertyId": property_id,
                "name": "Property Name",
                "status": "active"
            }
            
        except Exception as e:
            logger.error(f"Error getting property {property_id}: {e}")
            raise
    
    async def send_event(
        self,
        property_id: str,
        event_name: str,
        params: Dict[str, Any]
    ) -> Dict[str, Any]:
        """Send custom event to GA4"""
        try:
            logger.info(f"Sending event '{event_name}' to property {property_id}")
            
            # TODO: Implement actual event sending via Measurement Protocol
            # For now, log and return success
            logger.info(f"Event params: {params}")
            
            return {
                "success": True,
                "propertyId": property_id,
                "eventName": event_name,
                "timestamp": "2026-02-03T00:00:00Z"
            }
            
        except Exception as e:
            logger.error(f"Error sending event: {e}")
            raise
    
    async def run_report(
        self,
        property_id: str,
        dimensions: List[str],
        metrics: List[str],
        date_range: Dict[str, str]
    ) -> Dict[str, Any]:
        """Run custom GA4 report"""
        try:
            if not self.data_client:
                raise Exception("Analytics Data client not initialized")
            
            logger.info(f"Running report for property {property_id}")
            
            # Build request
            request = RunReportRequest(
                property=f"properties/{property_id}",
                dimensions=[Dimension(name=d) for d in dimensions],
                metrics=[Metric(name=m) for m in metrics],
                date_ranges=[DateRange(
                    start_date=date_range.get("startDate"),
                    end_date=date_range.get("endDate")
                )]
            )
            
            # TODO: Execute actual report
            # response = self.data_client.run_report(request)
            
            # For now, return sample data
            return {
                "propertyId": property_id,
                "dateRange": date_range,
                "dimensions": dimensions,
                "metrics": metrics,
                "rows": []
            }
            
        except Exception as e:
            logger.error(f"Error running report: {e}")
            raise
